runqemu: Add background launching support
Add support to the runqemu script to launch a program in the background.
This command can be specified by setting the QB_BACKGROUND_COMMAND
variable.

This launches the command in the background before starting QEMU.

This is useful for starting servers or services that QEMU or the guest
will connect to. This is useful for deploying certain images or
debugging.

Signed-off-by: Alistair Francis <alistair.francis@xilinx.com>