own-mirrors.bbclass: Change prepend to append
If BB_NO_NETWORK is set to 1, and local download directory is added as
PREMIRRORS_prepend in conf file, PREMIRRORS variable will have
SOURCE_MIRROR_URL as the first url. This causes build to fail as network
access is diabled. PREMIRRORS should be appended here so that a prepend
in other layers/conf files takes precedence.

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>