kernel.bbclass: Fix creating backup of compressed/vmlinux during initramfs
When building initramfs bundle along with fitimage option, there is a
bug that increases the size of fitimage. This happens only for platforms
which support compressed vmlinux creation ex: arm platforms

The increase of size in fitimage is because during initramfs creation,
the images under compressed directory are not backed up. So during the
do_bundle_initramfs step this will end up creating a vmlinux with rootfs.
In the subsequent steps when the fitimage runs it uses the vmlinux (with
ramfs) to create linux.bin. Hence causing the size increase in fitimage.

This patch addresses the issue by creating backup explicitly for vmlinux
under the arch/arm/boot/compressed/ directory. It will restore the
vmlinux after initramfs compile to have both vmlinux and
vmlinux.initramfs in the same location.

Reported-by: Michael Gill <gill@xilinx.com>
Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>