kernel: move image symlink to on target package postinstall
Instead of creating /boot/Image at build time, we can perform this action in
postinstall script only run on target. This allows actions such as kernel
upgrade to be performed, since there are no errors from overlapping files when
two kernel packages are installed on the target.

In a standard install of the kernel-image package to the target, there is no
difference between this updated technique and the build time method (and in the
past this symlink was done at install time versus build time)

   Linux qemuarm64 5.0.7-yocto-standard #1 SMP PREEMPT Wed May 1 17:11:29 UTC 2019 aarch64 GNU/Linux
   root@qemuarm64:~# ls -alF /boot
   drwxr-xr-x    2 root     root          1024 May  4 04:25 ./
   drwxr-xr-x   18 root     root          1024 May  4 04:23 ../
   lrwxrwxrwx    1 root     root            26 May  4 04:25 Image -> Image-5.0.7-yocto-standard
   -rw-r--r--    1 root     root      13093376 May  4 04:17 Image-5.0.7-yocto-standard

Note: there could be issues with read-only roots, but in these scenarios
      kernel image symlinks can either be created at install time or
      via the same mechanism that all firstboot operations use.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>