devtool: Getting rid of esdk check for devtool create-workspace
*THIS IS A TEMPORARY FIX*
Currently we are installing esdks onto an nfs mount point as part of
petalinux tools. Once an esdk is installed, a .devtoolbase file is
created, which gets rid of the devtool create-workspace command. This is
an issue as we have to do a create-workspace and point to a local
directory on host to use this 'shared esdk' flow and we can't just
delete the .devtoolbase file as doing so will get rid of other devtool
commands like build-sdk. This patch gets rid of the check for
.devtoolbase file when creating the create-workspace command so that it
is available in and outside of the esdk context.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>