devtool: Add build dir as basepath for devtool
Adding fixed_setup as true to register the other commands if basepath set
to the BUILDDIR and if in a petalinux environment. (checking the PETALINUX
variable.) Earlier we added fixed_setup as true if its not a basepath this
was breaking the yocto default behaviour.
NOTE: This was a temp fix for devtool

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>
Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>