gcc-9.2: Backport relocation patch fix
Without relocating PREFIX/EXEC_PREFIX the system can not do runtime
relocation for the path to the usr/lib/gcc directory, and other components.

While this is not a normal or supported use-case it does work in the upstream
gcc.  This is difficult to test with the regular OE SDKs, as it requires
running the components with the correct LD_LIBRARY_PATH and ld.so.

Without this update, gcc will typically not be able to find the gcc
provided include file for stddef.h and similar.  This is due to certain
relocations being based on the PREFIX and/or EXEC_PREFIX locations which
are hardcoded at compilation time.

(From OE-Core rev: b879fe730bc2cbce99704705cb53fa9ee958b311)

Upstream-Status: backport commit 5507707f0d36a612992758dabd58d68a72ab0cfc

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>