menuconfig.py: Fix the srctree path for devtool menuconfig command
Menuconfig creates a config fragment with changes and populates it under
sources/oe-local-filesdir.

When user uses devtool modify <srctree> option it will extract the source in
specified path.  For this option the devtool menuconfig does not generate the
devtool-fragment.cfg. It checks for the .config file in <workspace>/sources.

Fix the srctree path by checking 'S' variable instead of assuming
<workspaces>/sources path.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>