Add mechanism for user to append to same devtool fragment after user runs finish
In current devtool flow, if user runs devtool modify, menuconfig and finish, it
will create a devtool-fragment.cfg and append to SRC_URI of the recipe.

When a user runs the same flow multiple times, the devtool-fragment.cfg created
in previous iteration gets replaced with the new fragment created in the
current iteration. As a result, user can lose config changes made previously.

Provide menuconfig with an option -a or --allow-append that lets users to
continue append to previous iteration of devtool-fragment.cfg.  Ex. devtool
menuconfig linux-xlnx -a

By default, the devtool flow will replace the config fragment unless specified
with the -a option.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>