archiver.bbclass: Make do_deploy_archives a recursive dependency
To ensure that archives are captured for all dependencies of a typical
bitbake build we add do_deploy_archives to the list of recursive
dependencies of do_build. Without this, archives may be missed for
recipes such as gcc-source which do not create packages or populate a
sysroot.

do_deploy_archives is also added to the recursive dependencies of
do_populate_sdk so that all sources required for an SDK can be captured.

Signed-off-by: Paul Barker <pbarker@konsulko.com>

This patch has not been accepted upstream, but is required to
automaticly generate archiver outout for populate_sdk and
populate_sdk_ext builds.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>