mesa-gl: Allow mesa-gl to build without X11 enabled.
Trying to build mesa-gl without X11 enabled, results in numerous errors
about being incorrectly configured, such as:

  ERROR: Problem encountered: building dri drivers require at least one windowing system or classic osmesa

A rendering engine must be available, so per the error we enable osmesa
if X11 is not available.  Once 'osmesa' is enabled we also need to make
sure that swrast for DRI is also available.

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>