standard.py: add check in devtool finish command if menuconfig is run or not
By default for recipes that inherit the kernel class, devtool finish command
creates a config fragment and adds it to the SRC_URI. It overrides the config
fragment created by menuconfig.

If the user is running devtool menuconfig and wants to keep appending to the
same  file, then use the run_do_menuconfig option to disable devtool finish
command from creating and overriding the config fragment created by menuconfig
command.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>