standard.py: Handle srctree when in detached state
If user uses external source tree which is in detached state, devtool finish
command fails with the following error:
b.process.ExecutionError: Execution of
'git checkout (HEAD detached at xilinx-v2020.1)' failed with exit code 1:
/bin/sh: -c: line 0: syntax error near unexpected token git checkout (HEAD
detached at xilinx-v2020.1)'

The error occurs as the checkedout branch is not determined if there are no
ovveride_branches and the source tree is in detached state.

To fix the issue:
First, print a warning to let the user know the source tree is in detached
state and the resulting devtool behaviour might not be as expected.
Second, during devtool finish, set the checkedout branch variable when in
detached state to prevent a git checkout.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>